+++ src/core/constant.rs
@@ -1,1 +1,2 @@
+/// Default JSON settings file name.
 pub const DEFAULT_FILE_NAME: &str = "settings.json";